python_worker/requirements.txt
google-genai==1.3.0
pydantic==2.8.0
requests==2.32.3
python-dotenv==1.0.1
